Process environment via hook in wrap.go
This is how Cobra's API intends for flags to be modified before the
main command is run, and it lets us nicely return an error in case of
failure during environment processing.

BindConfig is now free of any side effects, besides a panic in case
that the config to be bound is not processable. An annotation stores
the environment variable to be read into each flag at a later stage.